DAL technically finished, it now saves data from the model, although it is not tested yet

+++ EelData.Model/Warning.cs
@@ -2,7 +2,7 @@
 {
     public class Warning
     {
-        public int Priority { get; set; }
+        public byte Priority { get; set; }
         public string Message { get; set; }
     }
 }
